Bench avro

+++ sample/sample.go
@@ -26,7 +26,7 @@ type Item struct {
 	ID        string
 	CreatedAt time.Time
 	Visible   bool
-	OwnerID   uint32
+	OwnerID   int32
 
 	Name        string
 	Description *string
@@ -60,7 +60,7 @@ func init() {
 			ID:        genName(35),
 			CreatedAt: time.Now().Truncate(time.Millisecond),
 			Visible:   rand.Intn(10) > 0,
-			OwnerID:   rand.Uint32(),
+			OwnerID:   int32(rand.Int()),
 			Name:      genName(50),
 			Price:     rand.Float64(),
 			Weight:    rand.Float32(),
